+++ R/utils.R
@@ -81,46 +81,6 @@ brewer_ramp <- function(n, plt, delete.lights = TRUE) {
 #'
 #' @description Deletes elements from the main attribute list.
 #'
-#' @details This function is deprecated; use \code{\link{delete_attr}} instead.
-#'
-#' @param attrList Attribute list.
-#' @param ids ID numbers to delete from the list.
-#'
-#' @return The updated attribute list.
-#'
-#' @export
-#' @keywords internal
-deleteAttr <- function(attrList, ids) {
-  .Deprecate_once(msg = paste0("`deleteAttr` is deprecated and will be removed",
-                               " in a future version of EpiModel; use",
-                               " `delete_attr` instead"))
-
-  if (!inherits(attrList, "list")) {
-    stop("attrList must be a list", call. = FALSE)
-  }
-
-  attr_length <- vapply(attrList, length, numeric(1))
-  expected_length <- length(attrList[["active"]])
-  wrong_length_attr <- names(attr_length)[attr_length != expected_length]
-
-  if (length(wrong_length_attr > 0)) {
-    stop(
-      "The following attributes do not have the correct number of elements: \n",
-      paste0(wrong_length_attr, collapse = ", "),
-      "\n\n", "Check if they are initialized when new nodes are created."
-    )
-  }
-
-  if (length(ids) > 0) {
-    attrList <- lapply(attrList, function(x) x[-ids])
-  }
-  return(attrList)
-}
-
-#' @title Delete Elements from Attribute List
-#'
-#' @description Deletes elements from the main attribute list.
-#'
 #' @param dat Either the \code{netsim_dat} class object passed through
 #'            \code{netsim} simulations, or the main data object passed through
 #'            \code{icm} simulations.